requirements.txt
altgraph==0.17
fbs==0.9.0
future==0.18.2
iso8601==0.1.13
macholib==1.14
pefile==2019.4.18
PyInstaller==3.4
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
PyYAML==5.3.1
